Greek Brumalia Season Events (SF.01)
Added two seasonal events that will fire for Greeks or people in the Roman Administration relating to the Byzantine Festival of Brumalia practiced in Cosntantinople and the Empire until the 13th century. The first event involves sending holiday cakes as gifts, and the second gives the player the ability to Outlaw Brumalia, which will decrease general opinion but increase the opinion of Temple Vassals and the Religious Head.